refactor: centralize datatypes unit tests
